Moving test util and rules into separate module
