+++ receiptrepository/build.gradle
@@ -13,4 +13,5 @@ dependencies {
     api project(':commonrepository')
 
     testImplementation "org.robolectric:robolectric:$robolectricVersion"
+    testImplementation project(':testutils')
 }
