+++ receiptui/build.gradle
@@ -17,10 +17,13 @@ dependencies {
     implementation "androidx.paging:paging-runtime:$pagingRuntimeVersion"
 
     testImplementation "org.robolectric:robolectric:$robolectricVersion"
+    testImplementation project(':testutils')
+
     androidTestImplementation "androidx.test:rules:$testRulesVersion"
     androidTestImplementation "androidx.test.espresso:espresso-contrib:$espressoVersion"
     androidTestImplementation "androidx.test.espresso:espresso-intents:$espressoVersion"
     androidTestImplementation "com.squareup.okhttp3:mockwebserver:$mockServerVersion"
     androidTestImplementation "com.squareup.leakcanary:leakcanary-android-instrumentation:$leakcanaryVersion"
     androidTestImplementation "com.squareup.leakcanary:leakcanary-support-fragment:$leakcanaryVersion"
+    androidTestImplementation project(':testutils')
 }
