+++ testutils/build.gradle
@@ -1,0 +1,8 @@
+apply from: "$rootProject.projectDir/android-library.gradle"
+description = 'Test Utils for Hyperwallet UI SDK'
+
+dependencies {
+    implementation "androidx.test:rules:$testRulesVersion"
+    implementation "androidx.test.espresso:espresso-core:$espressoVersion"
+    implementation "com.squareup.okhttp3:mockwebserver:$mockServerVersion"
+}
