+++ testutils/config/lint.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<lint>
+    <issue id="VectorPath" severity="ignore"/>
+    <issue id="RtlEnabled">
+        <ignore regexp="(AndroidManifest).xml"/>
+    </issue>
+</lint>
