+++ testutils/src/main/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<manifest
+    package="com.hyperwallet.android.ui.testutils"/>
