+++ transfermethodui/build.gradle
@@ -23,7 +23,9 @@ dependencies {
     androidTestImplementation "com.squareup.okhttp3:mockwebserver:$mockServerVersion"
     androidTestImplementation "com.squareup.leakcanary:leakcanary-android-instrumentation:$leakcanaryVersion"
     androidTestImplementation "com.squareup.leakcanary:leakcanary-support-fragment:$leakcanaryVersion"
+    androidTestImplementation project(":testutils")
 
     testImplementation "org.robolectric:robolectric:$robolectricVersion"
     testImplementation "com.squareup.okhttp3:mockwebserver:$mockWebServerVersion"
+    testImplementation project(':testutils')
 }
