--- transfermethodui/src/androidTest/java/com/hyperwallet/android/ui/transfermethod/util/RecyclerViewCountAssertion.java
@@ -1,30 +1,0 @@
-package com.hyperwallet.android.ui.transfermethod.util;
-
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.MatcherAssert.assertThat;
-
-import android.view.View;
-
-import androidx.recyclerview.widget.RecyclerView;
-import androidx.test.espresso.NoMatchingViewException;
-import androidx.test.espresso.ViewAssertion;
-
-public class RecyclerViewCountAssertion implements ViewAssertion {
-    private final int mCount;
-
-    public RecyclerViewCountAssertion(int count) {
-        this.mCount = count;
-    }
-
-    @Override
-    public void check(View view, NoMatchingViewException noViewFoundException) {
-        if (noViewFoundException != null) {
-            throw noViewFoundException;
-        }
-
-        RecyclerView recyclerView = (RecyclerView) view;
-        RecyclerView.Adapter adapter = recyclerView.getAdapter();
-
-        assertThat(adapter.getItemCount(), is(mCount));
-    }
-}
