+++ transfermethodui/src/test/java/com/hyperwallet/android/ui/transfermethod/TransferMethodUtilsTest.java
@@ -24,7 +24,7 @@ import com.hyperwallet.android.model.transfermethod.HyperwalletBankCard;
 import com.hyperwallet.android.model.transfermethod.HyperwalletTransferMethod;
 import com.hyperwallet.android.model.transfermethod.PayPalAccount;
 import com.hyperwallet.android.ui.R;
-import com.hyperwallet.android.ui.transfermethod.rule.HyperwalletExternalResourceManager;
+import com.hyperwallet.android.ui.testutils.rule.HyperwalletExternalResourceManager;
 import com.hyperwallet.android.ui.transfermethod.view.TransferMethodUtils;
 
 import org.hamcrest.CoreMatchers;
